+++ tools/data/resize_videos.py
@@ -20,7 +20,8 @@ def resize_videos(vid_item):
     # Change the output video extension to .mp4 if '--to-mp4' flag is set
     if args.to_mp4:
         vid_path = vid_path.split('.')
-        assert len(vid_path) == 2, f"Video path '{vid_path}' contain more than one dot"
+        assert len(vid_path) == 2, \
+            f"Video path '{vid_path}' contain more than one dot"
         vid_path = vid_path[0] + '.mp4'
     out_full_path = osp.join(args.out_dir, vid_path)
     dir_name = osp.dirname(vid_path)